pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fen_to_image"
version = "0.1.0"
description = "Convert chess FEN strings to board images"
authors = [{ name = "Théo Gueuret", email = "tgueuret@live.fr" }]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "pillow",
]

[tool.setuptools.packages.find]
where = ["."]
